improve compatability with mocha error rendering
When mocha renders an error message that supports rendering a diff it checks for a prefix that matches the regexp /^([^:]+): expected/ (see [1]). When matching messages are found the test following this prefix is removed and the diff is rendered in it's place. By using this pattern expect-to error messages are rendered much clearer by mocha.

[1] https://github.com/mochajs/mocha/blob/c0f9be244bff479e948f59d1bdb825a45c2cb40c/lib/reporters/base.js#L208-L209